incusd/instance/qemu: Add agent drive support for Windows

+++ internal/server/instance/drivers/driver_qemu.go
@@ -2976,6 +2976,7 @@ func (d *qemu) spiceCmdlineConfig() string {
 // inside the VM's config volume so that it can be restricted by quota.
 // Requires the instance be mounted before calling this function.
 func (d *qemu) generateConfigShare() error {
+	isWindows := d.isWindows()
 	configDrivePath := filepath.Join(d.Path(), "config")
 
 	// Create config drive dir if doesn't exist, if it does exist, leave it around so we don't regenerate all
@@ -2985,88 +2986,92 @@ func (d *qemu) generateConfigShare() error {
 		return err
 	}
 
-	// Add the VM agent.
-	agentSrcPath, _ := exec.LookPath("incus-agent")
-	if util.PathExists(os.Getenv("INCUS_AGENT_PATH")) {
-		// Install incus-agent script (loads from agent share).
-		agentFile, err := incusAgentLoader.ReadFile("agent-loader/incus-agent")
-		if err != nil {
-			return err
-		}
-
-		err = os.WriteFile(filepath.Join(configDrivePath, "incus-agent"), agentFile, 0o700)
-		if err != nil {
-			return err
-		}
-
-		// Legacy support.
-		_ = os.Remove(filepath.Join(configDrivePath, "lxd-agent"))
-		err = os.Symlink("incus-agent", filepath.Join(configDrivePath, "lxd-agent"))
-		if err != nil {
-			return err
-		}
-	} else if agentSrcPath != "" {
-		// Install agent into config drive dir if found.
-		agentSrcPath, err = filepath.EvalSymlinks(agentSrcPath)
-		if err != nil {
-			return err
-		}
-
-		agentSrcInfo, err := os.Stat(agentSrcPath)
-		if err != nil {
-			return fmt.Errorf("Failed getting info for incus-agent source %q: %w", agentSrcPath, err)
-		}
-
-		agentInstallPath := filepath.Join(configDrivePath, "incus-agent")
-		agentNeedsInstall := true
-
-		if util.PathExists(agentInstallPath) {
-			agentInstallInfo, err := os.Stat(agentInstallPath)
-			if err != nil {
-				return fmt.Errorf("Failed getting info for existing incus-agent install %q: %w", agentInstallPath, err)
-			}
-
-			if agentInstallInfo.ModTime() == agentSrcInfo.ModTime() && agentInstallInfo.Size() == agentSrcInfo.Size() {
-				agentNeedsInstall = false
-			}
-		}
-
-		// Only install the agent into config drive if the existing one is different to the source one.
-		// Otherwise we would end up copying it again and this can cause unnecessary snapshot usage.
-		if agentNeedsInstall {
-			d.logger.Debug("Installing incus-agent", logger.Ctx{"srcPath": agentSrcPath, "installPath": agentInstallPath})
-			err = internalUtil.FileCopy(agentSrcPath, agentInstallPath)
+	if !isWindows {
+		// Add the VM agent loader.
+		agentSrcPath, _ := exec.LookPath("incus-agent")
+		if util.PathExists(os.Getenv("INCUS_AGENT_PATH")) {
+			// Install incus-agent script (loads from agent share).
+			agentFile, err := incusAgentLoader.ReadFile("agent-loader/incus-agent")
 			if err != nil {
 				return err
 			}
 
-			err = os.Chmod(agentInstallPath, 0o500)
+			err = os.WriteFile(filepath.Join(configDrivePath, "incus-agent"), agentFile, 0o700)
 			if err != nil {
 				return err
 			}
 
-			err = os.Chown(agentInstallPath, 0, 0)
+			if !isWindows {
+				// Legacy support.
+				_ = os.Remove(filepath.Join(configDrivePath, "lxd-agent"))
+				err = os.Symlink("incus-agent", filepath.Join(configDrivePath, "lxd-agent"))
+				if err != nil {
+					return err
+				}
+			}
+		} else if agentSrcPath != "" {
+			// Install agent into config drive dir if found.
+			agentSrcPath, err = filepath.EvalSymlinks(agentSrcPath)
 			if err != nil {
 				return err
 			}
 
-			// Ensure we copy the source file's timestamps so they can be used for comparison later.
-			err = os.Chtimes(agentInstallPath, agentSrcInfo.ModTime(), agentSrcInfo.ModTime())
+			agentSrcInfo, err := os.Stat(agentSrcPath)
 			if err != nil {
-				return fmt.Errorf("Failed setting incus-agent timestamps: %w", err)
+				return fmt.Errorf("Failed getting info for incus-agent source %q: %w", agentSrcPath, err)
+			}
+
+			agentInstallPath := filepath.Join(configDrivePath, "incus-agent")
+			agentNeedsInstall := true
+
+			if util.PathExists(agentInstallPath) {
+				agentInstallInfo, err := os.Stat(agentInstallPath)
+				if err != nil {
+					return fmt.Errorf("Failed getting info for existing incus-agent install %q: %w", agentInstallPath, err)
+				}
+
+				if agentInstallInfo.ModTime().Equal(agentSrcInfo.ModTime()) && agentInstallInfo.Size() == agentSrcInfo.Size() {
+					agentNeedsInstall = false
+				}
+			}
+
+			// Only install the agent into config drive if the existing one is different to the source one.
+			// Otherwise we would end up copying it again and this can cause unnecessary snapshot usage.
+			if agentNeedsInstall {
+				d.logger.Debug("Installing incus-agent", logger.Ctx{"srcPath": agentSrcPath, "installPath": agentInstallPath})
+				err = internalUtil.FileCopy(agentSrcPath, agentInstallPath)
+				if err != nil {
+					return err
+				}
+
+				err = os.Chmod(agentInstallPath, 0o500)
+				if err != nil {
+					return err
+				}
+
+				err = os.Chown(agentInstallPath, 0, 0)
+				if err != nil {
+					return err
+				}
+
+				// Ensure we copy the source file's timestamps so they can be used for comparison later.
+				err = os.Chtimes(agentInstallPath, agentSrcInfo.ModTime(), agentSrcInfo.ModTime())
+				if err != nil {
+					return fmt.Errorf("Failed setting incus-agent timestamps: %w", err)
+				}
+			} else {
+				d.logger.Debug("Skipping incus-agent install as unchanged", logger.Ctx{"srcPath": agentSrcPath, "installPath": agentInstallPath})
+			}
+
+			// Legacy support.
+			_ = os.Remove(filepath.Join(configDrivePath, "lxd-agent"))
+			err = os.Symlink("incus-agent", filepath.Join(configDrivePath, "lxd-agent"))
+			if err != nil {
+				return err
 			}
 		} else {
-			d.logger.Debug("Skipping incus-agent install as unchanged", logger.Ctx{"srcPath": agentSrcPath, "installPath": agentInstallPath})
+			d.logger.Warn("incus-agent not found, skipping its inclusion in the VM config drive", logger.Ctx{"err": err})
 		}
-
-		// Legacy support.
-		_ = os.Remove(filepath.Join(configDrivePath, "lxd-agent"))
-		err = os.Symlink("incus-agent", filepath.Join(configDrivePath, "lxd-agent"))
-		if err != nil {
-			return err
-		}
-	} else {
-		d.logger.Warn("incus-agent not found, skipping its inclusion in the VM config drive", logger.Ctx{"err": err})
 	}
 
 	agentCert, agentKey, clientCert, _, err := d.generateAgentCert()
@@ -3089,63 +3094,65 @@ func (d *qemu) generateConfigShare() error {
 		return err
 	}
 
-	// Systemd units.
-	err = os.MkdirAll(filepath.Join(configDrivePath, "systemd"), 0o500)
-	if err != nil {
-		return err
-	}
+	if !isWindows {
+		// Systemd units.
+		err = os.MkdirAll(filepath.Join(configDrivePath, "systemd"), 0o500)
+		if err != nil {
+			return err
+		}
 
-	// Systemd unit for incus-agent. It ensures the incus-agent is copied from the shared filesystem before it is
-	// started. The service is triggered dynamically via udev rules when certain virtio-ports are detected,
-	// rather than being enabled at boot.
-	agentFile, err := incusAgentLoader.ReadFile("agent-loader/systemd/incus-agent.service")
-	if err != nil {
-		return err
-	}
+		// Systemd unit for incus-agent. It ensures the incus-agent is copied from the shared filesystem before it is
+		// started. The service is triggered dynamically via udev rules when certain virtio-ports are detected,
+		// rather than being enabled at boot.
+		agentFile, err := incusAgentLoader.ReadFile("agent-loader/systemd/incus-agent.service")
+		if err != nil {
+			return err
+		}
 
-	err = os.WriteFile(filepath.Join(configDrivePath, "systemd", "incus-agent.service"), agentFile, 0o400)
-	if err != nil {
-		return err
-	}
+		err = os.WriteFile(filepath.Join(configDrivePath, "systemd", "incus-agent.service"), agentFile, 0o400)
+		if err != nil {
+			return err
+		}
 
-	// Setup script for incus-agent that is executed by the incus-agent systemd unit before incus-agent is started.
-	// The script sets up a temporary mount point, copies data from the mount (including incus-agent binary),
-	// and then unmounts it. It also ensures appropriate permissions for the Incus agent's runtime directory.
-	agentFile, err = incusAgentLoader.ReadFile("agent-loader/incus-agent-setup")
-	if err != nil {
-		return err
-	}
+		// Setup script for incus-agent that is executed by the incus-agent systemd unit before incus-agent is started.
+		// The script sets up a temporary mount point, copies data from the mount (including incus-agent binary),
+		// and then unmounts it. It also ensures appropriate permissions for the Incus agent's runtime directory.
+		agentFile, err = incusAgentLoader.ReadFile("agent-loader/incus-agent-setup")
+		if err != nil {
+			return err
+		}
 
-	err = os.WriteFile(filepath.Join(configDrivePath, "systemd", "incus-agent-setup"), agentFile, 0o500)
-	if err != nil {
-		return err
-	}
+		err = os.WriteFile(filepath.Join(configDrivePath, "systemd", "incus-agent-setup"), agentFile, 0o500)
+		if err != nil {
+			return err
+		}
 
-	err = os.MkdirAll(filepath.Join(configDrivePath, "udev"), 0o500)
-	if err != nil {
-		return err
-	}
+		err = os.MkdirAll(filepath.Join(configDrivePath, "udev"), 0o500)
+		if err != nil {
+			return err
+		}
 
-	// Udev rules to start the incus-agent.service when QEMU serial devices (symlinks in virtio-ports) appear.
-	agentFile, err = incusAgentLoader.ReadFile("agent-loader/systemd/incus-agent.rules")
-	if err != nil {
-		return err
-	}
+		// Udev rules to start the incus-agent.service when QEMU serial devices (symlinks in virtio-ports) appear.
+		agentFile, err = incusAgentLoader.ReadFile("agent-loader/systemd/incus-agent.rules")
+		if err != nil {
+			return err
+		}
 
-	err = os.WriteFile(filepath.Join(configDrivePath, "udev", "99-incus-agent.rules"), agentFile, 0o400)
-	if err != nil {
-		return err
-	}
+		err = os.WriteFile(filepath.Join(configDrivePath, "udev", "99-incus-agent.rules"), agentFile, 0o400)
+		if err != nil {
+			return err
+		}
 
-	// Install script for manual installs.
-	agentFile, err = incusAgentLoader.ReadFile("agent-loader/install.sh")
-	if err != nil {
-		return err
-	}
+		// Install script for manual installs.
+		agentFile, err = incusAgentLoader.ReadFile("agent-loader/install.sh")
+		if err != nil {
+			return err
+		}
 
-	err = os.WriteFile(filepath.Join(configDrivePath, "install.sh"), agentFile, 0o700)
-	if err != nil {
-		return err
+		err = os.WriteFile(filepath.Join(configDrivePath, "install.sh"), agentFile, 0o700)
+		if err != nil {
+			return err
+		}
 	}
 
 	// Templated files.
@@ -3190,45 +3197,47 @@ func (d *qemu) generateConfigShare() error {
 		}
 	}
 
-	// Clear NICConfigDir to ensure that no leftover configuration is erroneously applied by the agent.
-	nicConfigPath := filepath.Join(configDrivePath, deviceConfig.NICConfigDir)
-	_ = os.RemoveAll(nicConfigPath)
-	err = os.MkdirAll(nicConfigPath, 0o500)
-	if err != nil {
-		return err
-	}
+	if !isWindows {
+		// Clear NICConfigDir to ensure that no leftover configuration is erroneously applied by the agent.
+		nicConfigPath := filepath.Join(configDrivePath, deviceConfig.NICConfigDir)
+		_ = os.RemoveAll(nicConfigPath)
+		err = os.MkdirAll(nicConfigPath, 0o500)
+		if err != nil {
+			return err
+		}
 
-	// Add the NIC config.
-	if util.IsTrue(d.expandedConfig["agent.nic_config"]) {
-		sortedDevices := d.expandedDevices.Sorted()
-		for _, entry := range sortedDevices {
-			if entry.Config["type"] != "nic" {
-				continue // Only keep NIC devices.
+		// Add the NIC config.
+		if util.IsTrue(d.expandedConfig["agent.nic_config"]) {
+			sortedDevices := d.expandedDevices.Sorted()
+			for _, entry := range sortedDevices {
+				if entry.Config["type"] != "nic" {
+					continue // Only keep NIC devices.
+				}
+
+				dev, err := d.FillNetworkDevice(entry.Name, entry.Config)
+				if err != nil {
+					return err
+				}
+
+				err = d.writeNICDevConfig(dev["mtu"], entry.Name, dev["name"], dev["hwaddr"])
+				if err != nil {
+					return fmt.Errorf("Failed writing NIC config for device %q: %w", entry.Name, err)
+				}
 			}
+		}
 
-			dev, err := d.FillNetworkDevice(entry.Name, entry.Config)
+		// Writing the connection info the config drive allows the agent to start /dev/incus very
+		// early. This is important for systemd services which want or require /dev/incus/sock.
+		connInfo, err := d.getAgentConnectionInfo()
+		if err != nil {
+			return err
+		}
+
+		if connInfo != nil {
+			err = d.saveConnectionInfo(connInfo)
 			if err != nil {
 				return err
 			}
-
-			err = d.writeNICDevConfig(dev["mtu"], entry.Name, dev["name"], dev["hwaddr"])
-			if err != nil {
-				return fmt.Errorf("Failed writing NIC config for device %q: %w", entry.Name, err)
-			}
-		}
-	}
-
-	// Writing the connection info the config drive allows the agent to start /dev/incus very
-	// early. This is important for systemd services which want or require /dev/incus/sock.
-	connInfo, err := d.getAgentConnectionInfo()
-	if err != nil {
-		return err
-	}
-
-	if connInfo != nil {
-		err = d.saveConnectionInfo(connInfo)
-		if err != nil {
-			return err
 		}
 	}
 
@@ -3840,16 +3849,18 @@ func (d *qemu) generateQemuConfig(machineDefinition string, cpuInfo *cpuTopology
 		bus.allocate(busFunctionGroupNone)
 	}
 
-	// Write the agent mount config.
-	agentMountJSON, err := json.Marshal(agentMounts)
-	if err != nil {
-		return nil, fmt.Errorf("Failed marshalling agent mounts to JSON: %w", err)
-	}
+	if !d.isWindows() {
+		// Write the agent mount config.
+		agentMountJSON, err := json.Marshal(agentMounts)
+		if err != nil {
+			return nil, fmt.Errorf("Failed marshalling agent mounts to JSON: %w", err)
+		}
 
-	agentMountFile := filepath.Join(d.Path(), "config", "agent-mounts.json")
-	err = os.WriteFile(agentMountFile, agentMountJSON, 0o400)
-	if err != nil {
-		return nil, fmt.Errorf("Failed writing agent mounts file: %w", err)
+		agentMountFile := filepath.Join(d.Path(), "config", "agent-mounts.json")
+		err = os.WriteFile(agentMountFile, agentMountJSON, 0o400)
+		if err != nil {
+			return nil, fmt.Errorf("Failed writing agent mounts file: %w", err)
+		}
 	}
 
 	// process any user-specified overrides
